+++ frontend/src/pages/SongDetailPage.tsx
@@ -1,25 +1,18 @@
-import { useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
-import CommentList from '@/features/comments/components/CommentList';
-import KillingPartInfo from '@/features/songs/components/KillingPartInfo';
-import KillingPartTrackList from '@/features/songs/components/KillingPartTrackList';
+import KillingPartInterface from '@/features/songs/components/KillingPartInterface';
+import { KillingPartInterfaceProvider } from '@/features/songs/components/KillingPartInterfaceProvider';
 import Thumbnail from '@/features/songs/components/Thumbnail';
 import { useGetSongDetail } from '@/features/songs/remotes/useGetSongDetail';
+import { VideoPlayerProvider } from '@/features/youtube/components/VideoPlayerProvider';
 import Youtube from '@/features/youtube/components/Youtube';
-import useVideoPlayerContext from '@/features/youtube/hooks/useVideoPlayerContext';
 import Flex from '@/shared/components/Flex';
 import Spacing from '@/shared/components/Spacing';
-import SRAlert from '@/shared/components/SRAlert';
 import SRHeading from '@/shared/components/SRHeading';
-import ToggleSwitch from '@/shared/components/ToggleSwitch/ToggleSwitch';
 
 const SongDetailPage = () => {
-  const { id: songId = '' } = useParams();
-  const { songDetail } = useGetSongDetail(Number(songId));
-
-  const [isRepeat, setIsRepeat] = useState(true);
-  const [killingRank, setKillingRank] = useState<number | null>(null);
+  const { id: songIdParams = '' } = useParams();
+  const { songDetail } = useGetSongDetail(Number(songIdParams));
 
   // useEffect(() => {
   //   if (!videoPlayer.current?.seekTo) return;
@@ -52,14 +45,9 @@ const SongDetailPage = () => {
   // }, [isRepeat, killingRank, videoPlayer.current, songDetail]);
 
   if (!songDetail) return;
-  const { killingParts, singer, title, songVideoUrl, albumCoverUrl } = songDetail;
-  const killingPart = killingParts.find((part) => part.rank === killingRank);
+  const { id: songId, killingParts, singer, title, songVideoUrl, albumCoverUrl } = songDetail;
 
   const videoId = songVideoUrl.replace('https://youtu.be/', '');
-
-  const toggleRepetition = () => {
-    setIsRepeat(!isRepeat);
-  };
 
   return (
     <Wrapper>
@@ -74,28 +62,13 @@ const SongDetailPage = () => {
         </Info>
       </SongInfoContainer>
       <Spacing direction="vertical" size={20} />
-      <Youtube videoId={videoId} />
-      <Spacing direction="vertical" size={16} />
-      <FlexContainer>
-        <TitleWrapper aria-label="Top 3 킬링파트 듣기">
-          <ItalicTitle aria-hidden="true">Top 3</ItalicTitle>
-          <NormalTitle aria-hidden="true"> Killing part</NormalTitle>
-        </TitleWrapper>
-        <SwitchWrapper>
-          <SwitchLabel htmlFor="repetition">구간 반복</SwitchLabel>
-          <ToggleSwitch
-            id="repetition"
-            on={toggleRepetition}
-            off={toggleRepetition}
-            defaultToggle={isRepeat}
-          />
-        </SwitchWrapper>
-      </FlexContainer>
-      <Spacing direction="vertical" size={16} />
-      <KillingPartTrackList killingParts={killingParts} />
-      <Spacing direction="vertical" size={10} />
-      {killingPart && <CommentList songId={songId} partId={killingParts[killingRank! - 1].id} />}
-      <SRAlert>{`${killingRank === 4 ? '전체' : `${killingRank}등 킬링파트`} 재생`}</SRAlert>
+      <VideoPlayerProvider>
+        <Youtube videoId={videoId} />
+        <Spacing direction="vertical" size={16} />
+        <KillingPartInterfaceProvider songId={songId}>
+          <KillingPartInterface killingParts={killingParts} />
+        </KillingPartInterfaceProvider>
+      </VideoPlayerProvider>
     </Wrapper>
   );
 };
@@ -144,39 +117,4 @@ const Singer = styled.div`
   @media (max-width: ${({ theme }) => theme.breakPoints.md}) {
     font-size: 16px;
   }
-`;
-
-const SwitchWrapper = styled.div`
-  display: flex;
-  column-gap: 8px;
-`;
-
-const SwitchLabel = styled.label`
-  font-size: 12px;
-  color: ${({ theme: { color } }) => color.white};
-`;
-
-const TitleWrapper = styled.div`
-  font-size: 22px;
-  font-weight: 700;
-  color: ${({ theme: { color } }) => color.white};
-
-  @media (max-width: ${({ theme }) => theme.breakPoints.md}) {
-    font-size: 18px;
-  }
-`;
-
-const ItalicTitle = styled.span`
-  font-style: italic;
-  color: ${({ theme: { color } }) => color.primary};
-`;
-
-const NormalTitle = styled.span`
-  color: ${({ theme: { color } }) => color.white};
-`;
-
-const FlexContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
 `;
